\name{normalizar.estandarizar}
\alias{normalizar.estandarizar}
\title{Normalize and standarize variables}
\usage{
normalizar.estandarizar(data)
}
\description{
Normalizes and standarizes the input variables.
}
\arguments{
\item{data}{A vector or data frame}
}
\value{
A numeric vector or data frame.
}
\examples{
dat <- c(1,3,4,5)
normalizar.estandarizar(dat)
}
